C++ implementation of the EqF to estimate bias.
