--- examples/ABC_EQF/CMakeLists.txt
@@ -1,14 +1,0 @@
-add_executable(ABC_EqF
-        main.cpp
-        EqF.cpp
-        State.cpp
-        Input.cpp
-        G.cpp
-        Direction.cpp
-        Measurements.cpp
-        utilities.cpp
-        runEQF_withcsv.h
-)
-
-target_link_libraries(ABC_EqF gtsam)
-
