+++ examples/CMakeLists.txt
@@ -17,4 +17,3 @@ if (NOT GTSAM_USE_BOOST_FEATURES)
 endif()
 
 gtsamAddExamplesGlob("*.cpp" "${excluded_examples}" "gtsam;${Boost_PROGRAM_OPTIONS_LIBRARY}" ${GTSAM_BUILD_EXAMPLES_ALWAYS})
-add_subdirectory(ABC_EQF)
